+++ modules/__tests__/Router-test.js
@@ -4,7 +4,6 @@ import { render, unmountComponentAtNode } from 'react-dom'
 import createHistory from 'history/lib/createMemoryHistory'
 import Route from '../Route'
 import Router from '../Router'
-import RoutingContext from '../RoutingContext'
 
 describe('Router', function () {
 
@@ -277,98 +276,38 @@ describe('Router', function () {
 
   })
 
-  describe('RoutingContext', function () {
-    it('applies custom RoutingContext', function (done) {
-      const Parent = ({ children }) => <span>parent:{children}</span>
-      const Child = () => <span>child</span>
-
-      class LabelWrapper extends Component {
-        constructor(props, context) {
-          super(props, context)
-          this.createElement = this.createElement.bind(this)
-        }
-
-        createElement(component, props) {
-          const { label, createElement } = this.props
-
-          return (
-            <span>
-              {label}-inner:{createElement(component, props)}
-            </span>
-          )
-        }
-
-        render() {
-          const { label, children } = this.props
-          const child = React.cloneElement(children, {
-            createElement: this.createElement
-          })
-
-          return (
-            <span>
-              {label}-outer:{child}
-            </span>
-          )
-        }
-      }
-
-      LabelWrapper.defaultProps = {
-        createElement: React.createElement
-      }
-
-      const CustomRoutingContext = props => (
-        <LabelWrapper label="m1">
-          <LabelWrapper label="m2">
-            <RoutingContext {...props} />
-          </LabelWrapper>
-        </LabelWrapper>
-      )
-
+  describe('render prop', function () {
+    it('renders with the render prop', function (done) {
       render((
         <Router
-          history={createHistory('/child')}
-          RoutingContext={CustomRoutingContext}
-        >
-          <Route path="/" component={Parent}>
-            <Route path="child" component={Child} />
-          </Route>
-        </Router>
+          history={createHistory('/')}
+          render={() => <div>test</div>}
+          routes={{ path: '/', component: Parent }}
+        />
       ), node, function () {
-        // Note that the nesting order is inverted for `createElement`, because
-        // the order of function application is outermost-first.
-        expect(node.textContent).toBe(
-          'm1-outer:m2-outer:m2-inner:m1-inner:parent:m2-inner:m1-inner:child'
-        )
+        expect(node.textContent).toBe('test')
         done()
       })
     })
 
-    it('passes router props to custom RoutingContext', function (done) {
+    it('passes router props to render prop', function (done) {
       const MyComponent = () => <div />
       const route = { path: '/', component: MyComponent }
 
-      const Wrapper = (
-        { routes, components, foo, RoutingContext, children }
-      ) => {
-        expect(routes).toEqual([ route ])
-        expect(components).toEqual([ MyComponent ])
-        expect(foo).toBe('bar')
-        expect(RoutingContext).toNotExist()
+      const assertProps = (props) => {
+        expect(props.routes).toEqual([ route ])
+        expect(props.components).toEqual([ MyComponent ])
+        expect(props.foo).toBe('bar')
+        expect(props.render).toNotExist()
         done()
-
-        return children
+        return <div/>
       }
-      const CustomRoutingContext = props => (
-        <Wrapper {...props}>
-          <RoutingContext {...props} />
-        </Wrapper>
-      )
 
       render((
         <Router
           history={createHistory('/')}
           routes={route}
-          RoutingContext={CustomRoutingContext}
+          render={assertProps}
           foo="bar"
         />
       ), node)
@@ -377,12 +316,12 @@ describe('Router', function () {
   })
 
   describe('async components', function () {
-    let componentSpy, RoutingSpy
+    let componentSpy, renderSpy
 
     beforeEach(function () {
       componentSpy = expect.createSpy()
 
-      RoutingSpy = ({ components }) => {
+      renderSpy = ({ components }) => {
         componentSpy(components)
         return <div />
       }
@@ -395,7 +334,7 @@ describe('Router', function () {
       }
 
       render((
-        <Router history={createHistory('/')} RoutingContext={RoutingSpy}>
+        <Router history={createHistory('/')} render={renderSpy}>
           <Route path="/" getComponent={getComponent} />
         </Router>
       ), node, function () {
@@ -415,7 +354,7 @@ describe('Router', function () {
       }
 
       render((
-        <Router history={createHistory('/')} RoutingContext={RoutingSpy}>
+        <Router history={createHistory('/')} render={renderSpy}>
           <Route path="/" getComponents={getComponents} />
         </Router>
       ), node, function () {